feat: add leaf hash logic using blake2b hint

+++ crates/cairo_runner/src/types.rs
@@ -1,5 +1,7 @@
 #[derive(Debug, Clone)]
 pub struct InputData {
-    pub header_bytes: Vec<u8>,
+    pub header_bytes: Vec<u32>,
     pub solution_indexes: Vec<u32>,
 }
+
+
